+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change Log
 
+### Upcoming
+
+- [#321](https://github.com/bcoe/yargs/pull/321) Custom config parsing function (@bcoe)
+
 ### v3.31.0 (2015/12/03 10:15 +07:00)
 
 - [#239](https://github.com/bcoe/yargs/pull/239) Pass argv to commands (@bcoe)